Giving the bot more commands!
I think the bot works okay so far but we need to find a new way for the bot to send messages other than interaction.response, as this does not seem like a good viable long term method of having an interaction with a user.

I'll research that later and try to implement it when I can.