Adding focus mode with no set end date, and more!
Updated the bot to now allow users to display ALL members in focus regardless of their usage of the bot. Also updated the bot to now include the ability to go into focus mode without necessarily having a set end date (bypassing the database altogether). This also means you don't get logged to the database if you use this, which may be preferable for my more paranoid people. :)