Updating study bot to include hourly posture and hydration reminders!
Created a reminder channel in the study fam discord server and the bot will now run a loop to post that every single hour.

Will need to update this eventually to include various different responses for the bot so it doesn't get too stale. Anyway, this will do for now!