Turning off self care reminders for now
GIving people a break from the constant spam of the channel. Trying to think of a better, less spammy way to do this with the same good effect but not make it so annoying.

(Though I understand that of course some of it will always be annoying, hence what its for. But it shouldn't be this bad. lol).