Auto Delete channel added!
Now the discord server has a channel that will have every message deleted after 24 hours from when it was sent.

I built this routine into the role and removal function and changed that function name to bot_routines instead. But to be honest I probably made this function way too complex and long. But hopefully the code should make sense. I will rework it and refactor it later, eventually.