Adding one more if statement
Adding a condition to check if there are users in focus that are not in the database before stating the following strings after the if statement. Because, if there are no users in the role not in the db, then we don't even need to bother doing that stuff. Thus an if statement works well here.