Fixing the bot once more
Apparently I did not include a possible scenario in which there is NOTHING in the database. Which I should have thought of, but figured it would never come up. lol. But I fixed that now with a little more messy code. I will work on a cleaner implementation eventually. (Famous last words) lol