Updating self care reminders
Cleaned up the code a lot. Completely bypassing the need for a database, which makes more sense anyway to do that. Removed almost a hundred lines of code in the main script. I kept the functions from the database script in tact though, mainly for a rainy day I guess. I will clean that up eventually though.