Fixing dr k content pings
Now the bot will ping the role every time carl posts in that channel, which could be abused... For some reason even though I enabled the message intents, it still can't actually read the contents of the message, so it's coming back as empty strings, causing my IF statements to fail. So I can't control for that really. I am sure there may be some kind of workaround... but honestly I don't care at this point. It will work for now.