Updating the bot to add more info to the txt files
Adding the functionality of adding the video title, id, and date the video was uploaded... well at least when carl bot posts in discord about it anyway. But close enough I guess. All that really matters is the date, not necessarily the time. The video title should be enough to look up a video but unfortunately if they ever change a video title after upload (which could happen often) then it will make the video unable to be found. However, luckily having the date and the video ID from Pytube should be enough to locate the video.