Adding the potential for food related commands
Might do more with this. I have some ideas. Like a command where you can get a random recipe to try to make. Or a command where you can type in some search terms and maybe tweak some parameters like "high protein" and it will provide a list of recipes. Or maybe you could just type a single phrase like "chicken soup" and it would return a list of recipes for Chicken Soup related stuff.

Not entirely sure what I wanna do with it yet. Seems like this might not be the best content for discord or at least out of my scope right now. But it's a cool thought experiment nonetheless.