Adding a random advice status & command
Using the "Advice API" we can now get the discord bot to give us a random piece of advice as its status every single time it boots up, as well as a discord command "get_random_advice" that you can use to send in a channel. Kind of like a magic 8 ball basically.