making non-member only content pings
Added some extra stuff so that we can control for private YT content and not send out pings or transcription threads of member content. (rip members).

We could make member specific ping roles, but eh, this will work at a small scale. But on a large scale you would need to address this.