Adding automatic image transcriptions
The way this works is that when the user uploads an image in their message as an attachment, it will save the image locally, OCR it, remove it, and then upload the contents of that text string to a thread that it creates in reply to the message.

There is another alternative, we could store the image text to a SQL db and instead allow users to retrieve it via a slash command, but this introduces a lot more complexity. I also would prefer it stored on discord's servers, not my own storage.

It may look a little uglier to do it this way, but for now, it'll do.